+++ examples/login/logind.lua
@@ -1,26 +1,17 @@
-local login = require "gamefw.loginserver"
+local login = require "snax.loginserver"
 local crypt = require "crypt"
 local skynet = require "skynet"
 
 local server = {
 	host = "127.0.0.1",
 	port = 8001,
+	multilogin = false,	-- disallow multilogin
 	name = "login_master",
 }
 
 local server_list = {}
 local user_online = {}
-
-local server_mt = {}
-server_mt.__index = server_mt
-
-function server_mt:kick(uid)
-	return skynet.call(self.address, "lua", "kick", self.name, uid)
-end
-
-function server_mt:login(uid, secret)
-	return skynet.call(self.address, "lua", "login", self.name, uid, secret)
-end
+local user_login = {}
 
 function server.auth_handler(token)
 	-- the token is base64(user)@base64(server):base64(password)
@@ -34,27 +25,31 @@ end
 
 function server.login_handler(server, uid, secret)
 	print(string.format("%s@%s is login, secret is %s", uid, server, crypt.hexencode(secret)))
-	local u = user_online[uid]
-	if u then
-		u:kick(uid)
-	end
-	assert(user_online[uid] == nil, "kick failed")
 	local gameserver = assert(server_list[server], "Unknown server")
-	local subid = gameserver:login(uid, secret)
-	user_online[uid] = gameserver
-	return tostring(subid)
+	-- only one can login, because disallow multilogin
+	local last = user_online[uid]
+	if last then
+		skynet.call(last.address, "lua", "kick", uid, last.subid)
+	end
+	if user_online[uid] then
+		error(string.format("user %s is already online", uid))
+	end
+
+	local subid = tostring(skynet.call(gameserver, "lua", "login", uid, secret))
+	user_online[uid] = { address = gameserver, subid = subid , server = server}
+	return subid
 end
 
 local CMD = {}
 
 function CMD.register_gate(server, address)
-	server_list[server] = setmetatable( { name = server, address = address }, server_mt )
+	server_list[server] = address
 end
 
-function CMD.logout(uid)
+function CMD.logout(uid, subid)
 	local u = user_online[uid]
 	if u then
-		print(string.format("%s@%s is logout", uid, u.name))
+		print(string.format("%s@%s is logout", uid, u.server))
 		user_online[uid] = nil
 	end
 end